Make user screens title layout consistent
