--- portal/src/graphql/adminapi/UserDetailCommandBar.module.scss
@@ -1,3 +1,0 @@
-.commandBar {
-  border-bottom: 1px solid #edebe9;
-}
